refactor responsive sidebar to separate component
